refactor(cli): move the loopback client and wire types into a shared api module
The TUI's HTTP+WebSocket client and event-frame decoding are not specific to
the interactive surface; a non-interactive client needs the same contract.
Move them out of `tui/` into `api/` and update the import paths. No behavior
change.
